created exception folder and moved some of the functions to their own files
