+++ flask_monitoringdashboard/__init__.py
@@ -18,7 +18,9 @@ from flask import Blueprint, g
 
 from flask_monitoringdashboard.core.config import Config, TelemetryConfig
 from flask_monitoringdashboard.core.logger import log
-from flask_monitoringdashboard.core.scoped_exception_logger import ScopedExceptionLogger
+from flask_monitoringdashboard.core.exceptions.scoped_exception_logger import (
+    ScopedExceptionLogger,
+)
 
 
 def loc():
