+++ flask_monitoringdashboard/core/profiler/outlier_profiler.py
@@ -8,7 +8,7 @@ from flask import request
 
 from flask_monitoringdashboard import config
 from flask_monitoringdashboard.core.cache import update_duration_cache, get_avg_endpoint
-from flask_monitoringdashboard.core.exception_logger import ExceptionLogger
+from flask_monitoringdashboard.core.exceptions.exception_logger import ExceptionLogger
 from flask_monitoringdashboard.core.logger import log
 from flask_monitoringdashboard.database import session_scope
 from flask_monitoringdashboard.database.outlier import add_outlier
